// src/app/interfaces/weatherApiResponse.interface.ts
export interface WeatherApiResponse {
    city: {
        id: number;
        name: string;
        coord: {
            lat: number;
            lon: number;
        };
        country: string;
        population: number;
        timezone: number;
    };
    list: {
        dt: number;
        main: {
            temp: number;
            feels_like: number;
            temp_min: number;
            temp_max: number;
            pressure: number;
            sea_level: number;
            grnd_level: number;
            humidity: number;
            temp_kf: number;
        };
        weather: {
            id: number;
            main: string;
            description: string;
            icon: string;
        }[];
        clouds: {
            all: number;
        };
        wind: {
            speed: number;
            deg: number;
        };
        visibility: number;
        pop: number;
        rain: {
            "3h": number;
        };
        sys: {
            pod: string;
        };
        dt_txt: string;
    }[];
}